// src/formateurs/dto/token.dto.ts
import { IsNotEmpty, IsEmail } from 'class-validator';
export class TokenFDto {
  @IsNotEmpty()
  id?: string;
  @IsNotEmpty()
  name: string;
  @IsNotEmpty()
  lastname: string;
  @IsNotEmpty()
  tel: number;
  @IsNotEmpty()
  isValid: boolean;
  @IsNotEmpty()
  @IsEmail()
  email: string;
  @IsNotEmpty()
  password: string;
  logo: string;
  spec: string;
  idR: string[];
}